Readonly procedures for causal cluster
